Update screenshots for new chrome version
The latest version of chrome triggered a bug somewhere so that
screenshots on mobile now have the wrong aspect ratio.
Upgrading webdriverio to version 8 would fix the issue,
however this then breaks compatibility with jest.

Since it is a bigger change to make the screenshot infrastructure
work again, this PR just adapts to the new screenshots (since they
are at least stable) so that we can deal with fixing the infrastructure
properly at a later stage.